+++ index.html
@@ -228,8 +228,8 @@
     </section>
   </template>
 
+  <script src="js/main.js" type="module"></script>
   <script src="js/functions.js"></script>
-  <script src="js/main.js"></script>
 </body>
 
 </html>
